dispatcher_engine/requirements.txt
aiosqlite==0.17.0
anyio==3.6.2
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.94.1
h11==0.14.0
idna==3.4
iso8601==1.1.0
passlib==1.7.4
psycopg2==2.9.5
pydantic==1.10.6
PyJWT==2.6.0
pypika-tortoise==0.1.6
python-multipart==0.0.6
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.6
starlette==0.26.1
tortoise-orm==0.19.3
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.0
